+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [4.0.3] 2022-6-20
+
+- Custom error type in axum request extractor [#945](https://github.com/async-graphql/async-graphql/pull/945)
+- Add nodes exposure on `ConectionType` so nesting through edges isn't always needed. [#952](https://github.com/async-graphql/async-graphql/pull/952)
+- Make email-validator optional [#950](https://github.com/async-graphql/async-graphql/pull/950)
+
 # [4.0.2] 2022-6-10
 
 - Expose `Edge::node` to allow better testing. [#933](https://github.com/async-graphql/async-graphql/pull/933)